new move folders, ref pass size in get_valid_heap
