+++ lib/Memory_Manager/init_methods/init_fp_copied_with_spif_func.cpp
